couple of remaining rel 8.8 issues (#2977)
* fix std::out_of_range exception when slicing an empty list

The starting point was a 0 constant mask, but it seems that works
through the compiler for other tests. However, in the p4c-1162.p4 the
test ends up creating an empty possibleSlicingPoints list (why? not
sure), and thus we check the size of the list before using it later in
the code.

P4C-1162  #assign #start-work #fixed fix std::out_of_range exception

* wordsmith error message

Make it gramatically correct and try to point out the actual field.
It is still not clear how to get to the field name that was turned into
a temporary ...

P4C-682  #assign #start-work #fixed wordsmith error message