Fix iface check in ptf_runner
On Debian (8?), /sbin/ifconfig outputs a ':' at the end of each
interface name. When building a list of interfaces in ptf_runner, we
were including this ':' and displaying an error message saying some
interfaces were missing even when all of them were present. This commit
change the regex matching to omit the trailing ':' when present (so it
works on both Debian and Ubuntu).