All uses of previous field must be able to reach the initialization point (#3053)
When doing metadata initialization, we ensure that none of the uses of the later field reach the
initialization point. At the same time, all the uses of the previous field must reach the
initialization point, because all those uses of the earlier field must necessarily happen before the
initialization of the container to the later field.