adds dumping table dependency graphs (#3091)
* adds dumping table dependency graphs

We serialize the .dot files for the dependency graph and publish them
in the manifest, so they can be displayed in p4i.

There are two points in the program where the graphs are dumped:
- the first time the graph is computed -- capturing the "initial"
  program graph. This is called 'program_graph.dot'
- every time that table placement is called. These are called
  'table_placement_X.dot', where X is invocation instance.

The original code dumped 3 graphs:
- a graph with all tables and every edge (including multiple edges
  that represent the same type of dependency)
- a "merged" graph that merges edges of the same type
- a simple graph that simplifies the dependencies into CONTROL and DATA.

For now we dump the most complete graph. with the following changes:
- edges are colored with the type of dependency rather than labeled
- table names use the external name if available, and names are not escaped
- print a legend with the edge colors as a subgraph

Potential optimizations:
- name the table_placement graphs in a more "intuitive" way
- merge edges of the same type, or label them with the actual sources
  of the dependency
- use different shapes for tables of different types
- rather than cond_XX, label the conditions with the actual expression
- refine the colors?
- place the legend in the right-hand side bottom corner (need to read
  the details in the dot manual!)

An additional "optimization" is to turn on verbose 1 and create-graphs
when using -g in the compiler driver. This allows the generation of
logs, metrics, and graphs when the user wants -g. Seems to save a bit
of typing.  More verbose logging can still be specified explicitly on
the command line.