Fix corner case with ActionPhvConstraints (#3103)
In COMPILER-823, we have a case where non-contiguous slices are written in an action, and their
sources are unallocated. `ActionPhvConstraints` enforces that the unallocated sources be allocated
together in the same container. However, there was a missing check to ensure that the write to the
container can be synthesized using a `deposit-field` instruction with a valid mask. This PR adds a
check for a valid mask in this case.