Create slice lists out of metadata fields (#3111)
* Create slice lists out of metadata fields

For metadata fields that have the maximum container bytes requirement set, we determine if the
maximum container bytes is the same as the minimum container bytes (the size of the field rounded up
to the nearest byte). If this condition holds true, the field effectively has the exact containers
requirement. This is expressed by creating a slice list containing all the slices of the metadata
fields. This allows for a fitting of a large supercluster in COMPILER-650.

* Remove an overly constrained condition in AllocatePHV

* Update XFails