Fixes to slicing to resolve Arista bugs (#3118)
* Fixes to slicing to resolve Arista bugs

The Arista test exposes a case where `SlicingIterator` exceeds the slicing threshold during the
allocation search (and after pre-slicing) for a large supercluster. At the point, we have at least
one supercluster instance that can fit in the container sizes available on Tofino. The further
slicing only allows us to explore allocations to smaller sized containers. Instead of throwing an
error when we exceed the slicing threshold, the compiler will now use the original supercluster
structure to allocate PHVs, assuming that further slicing of the supercluster is no longer possible.

* Fix generation of conditional constraints

Ordinarily, when we produce conditional constraints for multiple unallocated slices, we try to
allocate all those unallocated slices into the same container. However, if one of the unallocated
slices has the `no-pack` property, then these conditional constraints can never be satisfied. So,
for this case, we ensure that the `no-pack` field are allocated in their own container.

* Handle edge cases

- Prevent allocation when we have more than two no-pack conditional constraints.
- Prevent allocation when we have two no-pack conditional constraints plus other field slices to be
  conditionally allocated.
- Prevent allocation when we have two conditional PHV sources and one action data/constant source.