Fix hash packing for random number generator (#3136)
* Fix hash packing for random number generator

This PR fixes the case presented in the example program attached here. In this test, a 29-bit random
number is generated and assigned to a field. The destination of the random number should have the
some container size constraints as fields that source their value from the results of a hash
operation. Also, when deciding those fields, if the source of a set is a slice, we need to look at
the original expression to identify the `HashDist` or `RandomNumber` source.

* Fix an issue with pa_container_size

* Update XFail