Be smarter about clobbering/appending to logs. (#3142)
This fixes parser logging, wherein `parser.log` from previous runs of
the compiler was being appended instead of clobbered, resulting in
potential misinterpretation of the log.

The `bool append` parameter of `Logging::FileLog` and
`Logging::PassManager` is replaced with an enum of `APPEND`, `CREATE`,
and `AUTO`. With `AUTO`, the log file will be clobbered the first time
the compiler writes to it; subsequent writes will append.

Special casing for parser logs in `Logging::PassManager` is eliminated.
We now only generate log suffixes if the pass manager is constructed
with logging mode `CREATE`.