Tweaks & fixes to commit-msg. (#3180)
* Remove `feature` feature.

This removes an unused feature that intended to annotate the first line
of the commit message. There were a few minor bugs that could have been
easily fixed, this feature is being removed for a couple of reasons. It
appears to require branch names formatted as
`user/feature-name/p4c-####`, but no one is doing this. The feature also
risks pushing the first line of the commit message over the 70-character
mark without the user's knowledge.

* tweak

* Fix for empty commit messages.

An empty commit message is supposed to abort the commit, so only
introduce the JIRA stanza if the commit message is non-empty.

A commit message is non-empty if it has a line that does not begin with
'#' and contains a non-whitespace character.