add source input files in the manifest (#3236)
* add source input files in the manifest

Add a new node in the program to represent the compiled source files.
The node has the root source directory, includes, and defines.

Add --archive-source support: since we now publish source information
in the manifest, we also give the option to the user to package the
source in the archive.  This is disabled by default because sharing
the source should be an explicit decision.

Also in this PR, we change the build system to read the schema versions
directly from the Python files and generate macros that can be used in the
code.

* re-factor error and warning messages in table layout and MAU ASM output

Move error and warning messages to typed errors/warnings to cut redundancy.