Slight refactoring of `foreach_alloc` and friends (#3274)
* Slight refactoring of `foreach_alloc` and friends

`foreach_alloc` and related methods take in an `IR::BFN::Unit*`, but
only examine its type in most cases. With this change, these functions
now instead take in a `PHV::AllocContext*`, so that we no longer need to
thread IR nodes through our method calls. This makes it easier to use
this and related APIs.

For instance, `ClotInfo.get_overwrite_containers` calls these methods,
and with this change, it no longer needs to take in an IR node, and can
just pass `PHV::AllocContext::DEPARSER` to `foreach_alloc` internally.

* cpplint