Fix issue with bf-diags on JBay (#3285)
* Fix issue with bf-diags on JBay

The issue was that I was automatically assuming that any table strictly dominates the deparser for
that gress. However, it is possible that another table (which is up the immediate dominator tree for
the table in question) is the actual immediate dominator for the deparser. In this case, the
assumption that the original table in question is strictly dominates the deparser is not true, as
the table itself may be inhibited at runtime. This was causing bf-diags to fail for Tofino2.

This PR fixes the issue with the strictlyDominates method.

P4C-1800 #assign #start-work Fix issue with bf-diags on JBay

* Fix issues with basic_ipv4 for Tofino2

There were several issues with dark allocation, including the following:
- Assuming that Device::numStages() denotes the deparser. This PR changes this to
  a value based either on the maximum dependency chain value in the dependency graph (for logical
  stages) and to the maximum number of stages + 1 (for physical stages).
- Also, fixes the logic for translation of logical to physical stages.

P4C-1808 #bug-present #assign #assign #start-work #fixed #fixed

* Fix bug with table dependency graph calculation of max_min_stage

P4C-1800 #assign #start-work Fix bug with table dependency graph calculation of max_min_stage

* Fix bugs with FinalizeStageAllocation

P4C-1800 #assign #start-work Fix bugs with FinalizeStageAllocation

* Deparser stage needs to be 0

P4C-1800 #assign #start-work Deparser stage needs to be 0

* Fix issue with multi-pipe compilation for JBay

P4C-1800 #assign #start-work Fix issue with multi-pipe compilation for JBay

* Try to get over fitting issues

P4C-1800 #assign #start-work Try to get over fitting issues

* Correct the sorting order for live slices

P4C-1800 #assign #start-work Correct the sorting order for live slices