Fix gateway merging (#3299)
* Fix gateway merging when table violates happens_after conditions

Gateway merging has not been checking whether potential merge tables have their happens_after
conditions satisfied. This change moves the search for mergeable tables into the preorder on pipes,
so that the preorder may prune the list with happens_after relationships. `gateway_merge_choices`
collects the initial list of potentially mergeable tables. If a gateway has tables that it can be
merged with, but they are not placeable due to happens_after conditions, then the gateway is skipped
for placement until these tables are placeable. `try_place_table` now takes a list of potential
merges as an argument.

* Fixed gateway skipping that could cause TP to wedge.

When gateways have tables they can be merged with, but those mergeable tables are not yet placeable
due to happens after restrictions, we want to skip placing the gateway until a mergeable table is
available. This is done to save on the number of logical tables. Minor fix was required to make this
work correctly.