Renamed happens_after/before_maps to more descriptive names. (#3310)
* Renamed happens_after/before_maps to more descriptive names.

The various happens_after/before_maps calculated in the table dependence graph were poorly named,
leading to possible misues. This commit renames them and updates table_placement.cpp to reflect
these changes, improving readability. However, table placement is still skipping tables based on
physical (stage) happens_after, as changing it to skip based on logical happens_after caused table
wedging.

* Further updates to names and uses.

Table placement now uses the logical map both for pruning the mergeable tables list and for deciding
to skip tables that are in the work list. Changed happens_after/before_map to
happens_after/before_work_map, to indicate that these maps are not to be used. Perhaps it's best to
privatize these variables, although that would require a more significant refactor to allow the
Inspector that builds the depedence graph access to these variables. Switched parde/decaf.cpp to use
the new name.

* Skip gateway in check of mergeable tables.

Gateway merging should not try to merge with tables that do not have all of the tables in logical
happens_after placed. However, this will always include the gateway itself, which we know will be
placed if the table is merged. Thus, we skip the gateway when performing this check. This change
should fix problems with table placement wedging.

* Added a gtest to check that physical/logical dependences are correct

* Added gtest to check that physical and logical deps are correct.