Fix tracking of extracts in CLOT allocator (#3340)
* Fix tracking of extracts in CLOT allocator.

This fixes a bug reported by @milad14000.

A field is CLOT-eligible only when it is extracted in a single state. However, the CLOT allocator was only tracking packet-sourced extracts. As a result, the implementation of CLOT eligibility was incorrect.

For example, in test case `p4_16/stf/extract_const.p4`, the field `hdr.h2.f` is not CLOT-eligible: it is extracted from the packet in state `h2b`, and from a constant in state `h2a`. But because the CLOT allocator only tracked packet-sourced extracts, the constant-sourced extract in state `h2a` was ignored, and the field was incorrectly deemed CLOT-eligible.