Fix generation of standard P4Info for TNA programs
It was broken after recent action selector changes. We now generate the
standard P4Info message for all the p4-tests TNA programs as part of the
testsuite to avoid that kind of issue in the future.

This commit also fixes a couple of smaller issues in teh BF-RT code. In
particular the "max group size" data field for action selector tables
always defaulted to 120 instead of using the value provided by the P4
programmer.

P4C-1988 #fixed #fixed