Transfer pa pragmas to alias destinations (#3490)
When two fields are aliased with each other, all uses of one of them (the alias source) are replaced
by the corresponding uses of the other (the alias destination). This replacement ensures that the
alias destination now reflects the union of the constraints of both the aliased fields.

However, some constraints are specified by the `pa_*` class of pragmas. Therefore, these constraints
should also be transferred to the alias destination fields, otherwise those constraints will no
longer be respected. This PR implements this change.

Note that the exception to this rule is the `pa_no_init` pragma because the live range of the alias
destination may well be very different from the live range of the alias source; and we cannot safely
shrink the live range from first use to last use for the alias destination.