Fix IR::MAU::Table::visit_children (#3332)
* Rewrote `MAU::Table::visit_children`.

This rewrites `MAU::Table::visit_children` so that the first
control-flow path through the `Table` is visited with `v`. Subsequent
paths are visited with clones of `v`'s initial state; the result is then
merged into `v`.

The hope is that this rewrite is clearer and better commented than its
predecessor, and is less buggy. The old code, for instance, visited all
paths through the `Table` with a clone of `v` that gets merged back into
`v`, which resulted in a phantom path that bypassed the table
completely.

* Fix: missing ampersands

* Use templates to reduce duplicated code.

I've discovered `#emit`...no longer need to bump p4c.

* Move some implementation code out of mau.def.

* Don't visit actions multiple times.

* Don't visit action chains multiple times.

* Preserve visit order.

This brings the order of visiting closer to the original code. Some
visitors appear to be sensitive to the order in which branches and
"next" entries are visited.

* Maybe this time?

* Be less timid about changing visit order.

Let's see what breaks.

* Revert "Be less timid about changing visit order."

This reverts commit 29d643a4429d713b048cbbd09b2806bdfc5f8d32.

* Was merging into v too early.

* Fix visiting of gateway payload and match table.

Unclear what to do if the gateway payload is unused and there is no
match table. Based on `Table::validate`, this can arise during table
placement. In these cases, the table represents a dead execution path,
but there doesn't appear to be a way to communicate dead execution paths
to the caller of `visit_children`. This commit treats these cases by
letting the control flow fall through to the `attached` table.

* Rescued a comment from #3370.

* Address comment.

Made it a `BUG` if we visit a table with neither a gateway payload nor a
match table.

* Slightly better treatment of dummy tables.

* Officially support placeholder-style tables.

Tables without a gateway payload and without a match table are now
considered to be no-ops, rather than dead execution paths.

* Updated xfails.

* Fix long branch props setting