Fix another corner case in ActionPhvConstraints (#3559)
* Fix another corner case in ActionPhvConstraints

When two different slices alive at the same time in a container are written using the same slice,
the source slice should count as two sources (one source for each destination slice; also extendable
to counting as n sources for n destination slices). This constraint was not captured by
`ActionPhvConstraints` and led to a failure for Arista. There are two different cases to be dealt
with here though: one where an unallocated source is used multiple times, and another where an
allocated source is used multiple times. This PR introduces code to fix both cases.

* Fix conditions

Need to explicitly only check for cases, where we have a double use for unallocated slices.
Otherwise, the unallocated slices could be packed together into a single container.