Constraint violation due to incorrect decision (#3533)
* Constraint violation due to incorrect decision

PHV allocation avoids container conflicts to give table placement the best chance of fitting
programs in the available number of stages. However, this creates a problem, when header fields that
must go into the same container are written by different tables in the same stage (in the container
conflicts-free table placement pass).

For such cases, we introduce the concept of `sliceable` superclusters. Nonsliceable superclusters
cannot be broken down into superclusters of smaller width, either because they are already at the
smallest available container width (8b) or they have some slice lists containing fields with
`no-split` constraints. This PR adds the ability to determine when a supercluster is nonsliceable;
and for those cases, `ActionPhvConstraints` allows container conflicts for fields packed together in
the same slice list.

This fixes an issue that Arista was running into with regards to PHV allocation.

* Address review comments

P4C-2058 #assign #start-work Address review comments