rename ^bridged_* to __bridged_* (#3588)
* rename ^bridged_* to __bridged_*

One more chapter in the naming saga: it seems that we automatically
strip the ^ characters from the names, thus creating two different
names in context.json for fields that are published for phv containers
and the same fields published as flexible_headers. The names are
prefixed with `__bfp4c_` to avoid clashes with user names.

For good measure, we declare these constant strings in a header and
use the constants throughout the compiler backend, rather than
the string itself. Makes for much easier renaming!

For usability, adjust some of the warning produced on switch to reduce
redundancy.

And for good measure, get rid of some compile warnings too!

P4C-2140 #assign #start-work rename ^bridged_* to __bridged_*