multipipe output and default bf-rt-schema argument (#3610)
* add --bf-rt-schema <output_dir>/bfrt.json by default if compiling P4-16
* add --no-bf-rt-schema flag and support for setting compiler arguments for tests

Setting the BF-RT schema generation by default works for most
programs, except when it doesn't (e.g. hwlearn* for Tofino2). And in
particular, there may be cases when we simply do not want to generate
it.  Therefore we add a new option, --no-bf-rt-schema to disable the
generation.

Also add support to runtest to read some compiler flags from the
CTEST_P4C_ARGS environment variable. This allows us to add arguments
to tests after they've been added, so that we don't need to exclude
and add tests manually.

* fix multipipe output generation when not verbose

Ensure that the manifest contains all the pipes even if no files
are produced by the compiler. The driver would run only if the
manifest has multiple pipes, and since we're no longer publishing
context.json in the case of a successful compilation, we were no
longer generating the second pipe.

Register the log files in the manifest in FileLog regardless of the
append flag. In the two pipe case, the file logger does not
distinguish by pipes, so logs for the second pipe would not be
registered in the manifest. To avoid duplicate records, we keep
the files in the manifest as a set.

* bfruntime fixes for Tofino2

Add BF-RT generation support for different bit types and dleft-hash
match type.

Also invoke the bf-runtime verifier only when the compiler runs.

P4C-1750  #assign #start-work #fixed #fixed

P4C-2160 #assign #start-work add --bf-rt-schema <output_dir>/bfrt.json by default if compiling P4-16