add support for defining targets outside the schema files (#3689)
* add support for defining targets outside the schema files

We abstract away the targets enumerated in the schema files, so that
we do not tip our hand on future targets.

The design is to create a file compiler-interfaces/schemas/targets.py
that defines the list of targets (and p4 architectures supported). If
the file is present, use it. Otherwise, use a predefined version of
the list with only the targets that are exposed.

P4C-2147  #assign #start-work add support for defining targets outside the schema files