P4C-2438: exit compiler driver cleanly if --help-pragmas requested
There is no need to run any other passes except the compiler with
the appropriate flags if invoked with --help-pragmas. So we force
running just the 'compiler' command and bail out early.

In the same shot, we need to treat ir_to_json in a similar manner,
but since there are multiple commands to be run, we just avoid
set a flag and avoid running verifiers.

P4C-2438 #assign #start-work P4C-2438: exit compiler driver cleanly if --help-pragmas requested