Fleshed out MultipleApply2, replaced MultipleApply (#3940)
* TablePlacement handle Table in mulitple sequences

* Check topological invariant for table applications

This introduces a `CheckTopologicalTables` pass to `MultipleApply2` for
checking that tables are applied in the same order on all paths through
the program. This is done by building a control-flow graph for tables;
any cycles in the graph will correspond to a set of tables applied in an
inconsistent order.

Also copied `MutuallyExclusiveApplies` and `DefaultNext` from
`MultipleApply` to `MultipleApply2`. The `CheckTopologicalTables` pass
assumes that tables are not applied more than once on any path.

The `MultipleApply` gtests have been updated to test `MultipleApply2`
instead.

* Replaced MultipleApply with MultipleApply2.

MultipleApply is dead. Long live MultipleApply!

* Fleshed out CheckStaticNextTable.

CheckStaticNextTable now handles tables not derived from P4 source-level
tables.

* Address a review comment from Chris.

* Tweaked an error message.

* Only mark duplicates after checking.

This addresses a review comment from @ChrisDodd.

* Tweaked CheckStaticNextTable::equiv_gateway

It turns out the `IR::Operation_Relation` case was redundant with
`IR::Operation_Binary`, since the former is a subtype of the latter.

Added deeper equivalence checking for unary and ternary operations.

* Fixed bug in DeduplicateTables

* Also de-duplicate TableSeq objects.

* Improved usefulness of a couple error messages

* Can't merge gateway with table in mulitple seqs

* Updated XFAILs for changed error messages

Also made some error messages consistent.

* Fix: don't deduplicate across threads

Table placement doesn't like it when things (e.g., empty TableSeqs) are
reused between gresses.

PackFlexibleHeaders runs MultipleApply on oddly formed IRs that have
multiple threads for the same gress. Deduplicating across these threads
causes DefaultNext to fail, so we don't de-duplicate across threads,
either.

* XFailing Arista's bare-metal profile

Not fitting PHVs. We'll kick this over to Mike Wang to figure out when
we give him the next drop.

* Address Mike's review comment

Co-authored-by: Chris Dodd <cdodd@acm.org>