P4C-2540: handle upcast in parser state assignment (#4039)
This PR fixes a bug in computation of valid_container_range of PHV::FieldSlice. As long as other parts of the code respect the constraint enforced by valid_container_range, the result should be correct. In this case, a unit test that only covers the corner case that triggers the bug is helpful:
82161c5

I have added a stf test as well. Stf tests surely have more testing coverage but (1) generally more coverage on the exercised path but less coverage on the specific issue. (2) it does not precisely describe the problem, may be hard to maintain later. (3) passing the test requires all components of the compiler to be correct, e.g. phv allocation respect container_valid_range, assembly of matching fields that are split across multiple containers are generated accrodingly, parser extract with the right offset and zero_inits are set for partial metadata. It is good and easy to write, but I would recommend that we write more short and precise unit tests that exercise more states of the issue.

Co-authored-by: Yumin Xia <yuminxia@intel.com>