P4C-2610: Add always-run tables to the IR (#4089)
* Finished a first cut at the implementation.

Insertion constraints are just a pair of sets of tables: one set for
tables that come before the insertion, and another for tables that come
after.

Added a gtest to kick the tires of the implementation. To aid with this,
table flow graphs have been augmented with support for finding dominator
sets, computed lazily. Also augmented the gtests for table flow graphs
to test these dominator sets.

* Addressed Evan's review

Co-authored-by: Jed Liu <jliu@barefootnetworks.com>