Fix environment variables in jarvis. (#4276)
* Fix environment variables.

It turns out the only reliable way to set environment variables is to
put them into the Dockerfile. The entry-point script isn't a good place,
since that is only executed in the terminal where the container is
started. Neither are bashrc nor bash_profile good places, because these
only execute when bash is started; if you run a command directly in a
container, for example, then neither bashrc nor bash_profile are
sourced.

[no-jira]

* Install distcc for all containers.

[no-jira]

* Tweak ccache config to account for upstream change

[no-jira]

Co-authored-by: Jed Liu <jliu@barefootnetworks.com>