P4C-611: Return an error when a P4 program try to use a non PHV field as the second argument of a subtraction (#4354)
* P4C-611: Return an error when a P4 program try to use a non PHV field as the second
argument of a subtraction. This use to fail in the assembler with
"error: src2 must be phv register". This modification try to help the
programmer finding the problem by referencing the P4 code part causing
the compile error.

P4C-611 #assign #start-work Return an error when a P4 program try to use a non PHV field as the second

* P4C-611: Only apply this check on the last operand

P4C-611 #assign #start-work Only apply this check on the last operand

* P4C-611: Handle the sizeInBytes and sizeInBits primitive.

P4C-611 #assign #start-work Handle the sizeInBytes and sizeInBits primitive.