P4C-2993: Code de-duplication, other improvements (#4427)
This is primarily a refactoring to avoid code duplication between
`output_jbay_field_dictionary` and `output_jbay_field_dictionary_slice`
by introducing a `output_jbay_field_dictionary_helper` that implements
the common control functionality. The other two functions now call this
helper and provide it with callbacks that implement the required
functionality for writing and finishing non-CLOT chunks, and for
handling CLOT chunks.

This commit also changes the use of `check_chunk`: we now check that
the current chunk is in range only when we are about to use it. This
avoids the edge case where we fail after having used the last available
chunk but before we attempt to use the first out-of-range chunk.

Additionally, calls to `tof2lab44_workaround` are now embedded into
`check_chunk` so that we skip bad chunk indices immediately before doing
the actual check.

P4C-2993 #assign #start-work Code de-duplication, other improvements

Co-authored-by: Jed Liu <jliu@barefootnetworks.com>