P4C-3168: Non-deterministic compilation results  (#4591)
* P4C-3168: Not ordered maps/sets were replaced with the ordered version.

This commits provides changes in the mau/phv code and it is using the
ordered version of map/set to preserve the insertion order. This is needed
because the mentioned containers were using a pointer as a key which means
(possible) different addresses between runs.
This can cause different order during the iteration over the container.

P4C-3168 #assign #start-work Not ordered maps/sets were replaced with the ordered version.

* P4C-3168: Changes based on the review:

* Some maps/sets are converted back to non-ordered version
* Typo in comment.

P4C-3168 #assign #start-work Changes based on the review:

* P4C-3168: The need_defer set is used inside the for cycle.

P4C-3168 #assign #start-work The need_defer set is used inside the for cycle.

* P4C-3168: Cleanup of maps/sets which don't need to be ordered.

P4C-3168 #assign #start-work Cleanup of maps/sets which don't need to be ordered.