P4C-3142: Error with cast operator in action analysis (support of the concat operator in keys) (#4685)
The order of the SimplifyKey and ElimCasts passes has been swapped.
The policy of the key simplification has been updated. Now, only a slice of a path expression is considered a simple key (apart from other policies). Before, a slice of anything was considered as simple keys, which was the issue.
The ElimCasts pass has been updated to not process key element's so that it doesn't interfere with the SimplifyKey pass.
As a result, the concatenation within a key element is simplified and then processed with the ElimCasts pass, which further slices it.