Don't assume Barefoot-specific options in primitive converters. (#4719)
Don't assume we have Barefoot-specific options, in case the current
compiler context is not a Barefoot-specific one. This occurs when the
Barefoot compiler is linked into an executable, but the user elects to
use a non-Barefoot backend with non-Barefoot options. In this scenario,
primitive converters that are statically registered by the Barefoot
compiler are still invoked.

Co-authored-by: Jed Liu <jliu@barefootnetworks.com>