P4C-3279: Fixed preservation of is_generated_compiler flag when merging a gateway table with a match table (#4772)
When merging a gateway table with a match table, the is_compiler_generated flag is lost. The reason is that the TransformTables::merge_match_and_gateway method does not preserve it while merging.
The consequence is that the table does not have the p4_hidden flag set in the context.json.

Added also preservation of the created_during_tp and has_dark_init flags.