P4C-3480: Fixing case where the port metadata was implicitly advanced. (#4890)
The compiler was implicitly adding a "pkt.advance" statement in the
parser when no port metadata field was extracted. The bit size computed
was incorrect and was in fact equal to the number of bits to advance.
This was not in-line with the extern signature
"void advance(in bit<32> sizeInBits)". The compiler was updated to
always use 32 for the number of bit used to carry the size to advance.

P4C-3480 #assign carl.geoffrion@intel.com