Use ccache when building in Jenkins (#5039)
When building in Jenkins, mount `~/.ccache` as `/root/.ccache` in the
Docker container.

Because `docker build` doesn't support mounting of external directories,
the Docker build is split into two stages. The first stage uses `docker
build` to install dependencies and set up the build environment. The
second stage uses `docker run` to mount `~/.ccache` and do the actual
build. Once this is done, we create an image out of the container's
final state.

[no-jira]

Co-authored-by: Jed Liu <jliu@barefootnetworks.com>