Place a register's write-after-read statements from an apply block into one action (#5044)
The compilation occasionally failed when a user tried to implement write-after-read behavior on a register array using the read and write methods directly in an apply block of a control block (i.e. outside an action). A sample of such problematic P4 code follows:

control SampleControl(...) {
    Register<value_t, index_t>(REG_SIZE) reg;

    apply {
        value_out = reg.read(index);
        reg.write(index, value_in);
    }
}

Depending on the rest of the user's P4 code, the compiler occasionally failed during the table placement phase.

The introduced improvement correctly detects write-after-read behavior in the context described above given that the call of the read method on a register array is directly followed by the call of the write method of the same register array as shown in the sample P4 code above.