P4C-3659: set non-zero source address for csum verif in Tof1 (#5073)
Tofino 1 requires that the last extractor be enabled when performing
checksum validation to work around a bug. Source address validation is
performed on this extractor.

To prevent source extract errors when buf_req = 0 (corresponding to
states with no extracts), set the source address to the "version"
section of the source range. The version is in bytes 60-63 and is always
valid.

P4C-3659 #assign glen.gibb@intel.com