Fixed parser match register handling in assembler (#5085)
In particular cases, the parser performs a wrong transition based on the select clause. It's quite difficult to reproduce such behavior with another P4 code; I haven't found a rule how to detect such situation.

The root cause of the issue is that match registers are handled in a wrong way in the code. Instead of treating load.save as a bit vector expressing occupancy of match registers, it is being treated as a logical value.

The proposed fix corrects the code and adds an STF test.