P4C-3665: Added support of the RegisterParam extern (#5101)
The compiler was extended with the support of register parameters. They are run-time configurable values, which can be used within register actions. Register parameters are represented by a RegisterParam extern with the following definition:

extern RegisterParam<T> {
    RegisterParam(T initial_value);
    T read();
}

A register parameter is equipped with an initial value, which is returned if the register parameter has not been configured.

An example of the usage of a register parameter follows:

Register<bit<32>, bit<10>>(1024) reg1;

RegisterParam<bit<32>>(2000) reg1param1;

RegisterAction<bit<32>, bit<10>, bit<32>>(reg1) reg1action1 = {
  void apply(inout bit<32> value, out bit<32> read_value) {
    if (hdr.data.f1 - value > reg1param1.read())
      value = hdr.data.f1;
  }
};

The use of a register parameter is limited to register actions bound to a single register array.

Register parameters can be used in the following contexts within a register action:

* To be returned from register actions
* To be stored to the corresponding register array
* To be used in comparisons in register actions

There is a limited number of slots that can be used to store register parameters. These slots are also used to store constants used in register actions. The compiler outputs an error if the capacity is exceeded.